refactor: reorganize test view files
