+++ src/ui/components/Library/Team/View/Tests/TestDetailsPanel/TestDetailsPanel.module.css
