+++ src/ui/components/Library/Team/View/Tests/TestDetailsPanel/TestDetailsPanel.tsx
@@ -6,7 +6,7 @@ import { TestSuitePanelMessage } from "../../TestSuitePanelMessage";
 import { useTest } from "../hooks/useTest";
 import { TestContext } from "../TestContextRoot";
 import { TestDetails } from "./TestDetails";
-import styles from "./TestOverviewContent.module.css";
+import styles from "./TestDetailsPanel.module.css";
 
 function SelectTestMessage({ error }: { error?: boolean }) {
   return error ? (
@@ -20,7 +20,7 @@ function SelectTestMessage({ error }: { error?: boolean }) {
   );
 }
 
-export function TestOverviewContent() {
+export function TestDetailsPanel() {
   const { testId, tests } = useContext(TestContext);
 
   let children = null;
