+++ src/ui/components/Library/Team/View/Tests/TestListPanel/TestListItem.tsx
@@ -3,9 +3,9 @@ import { useContext, useMemo } from "react";
 import { Test } from "shared/test-suites/TestRun";
 import Icon from "ui/components/shared/Icon";
 
-import HighlightedText from "../TestRuns/HighlightedText";
-import { TestRunLibraryRow } from "../TestRuns/TestRunLibraryRow";
-import { TestContext } from "./TestContextRoot";
+import HighlightedText from "../../HighlightedText";
+import { TestRunLibraryRow } from "../../TestRuns/TestRunLibraryRow";
+import { TestContext } from "../TestContextRoot";
 import styles from "./TestListItem.module.css";
 
 function Status({ test }: { test: Test }) {
