+++ src/ui/components/Library/Team/View/Tests/TestListPanel/TestListPanel.tsx
@@ -4,14 +4,14 @@ import { ContextMenuItem, useContextMenu } from "use-context-menu";
 import LibraryDropdownTrigger from "ui/components/Library/LibraryDropdownTrigger";
 import { LibrarySpinner } from "ui/components/Library/LibrarySpinner";
 
-import { TeamContext } from "../../TeamContextRoot";
-import { FilterField } from "../TestRuns/FilterField";
+import { TeamContext } from "../../../TeamContextRoot";
+import { FilterField } from "../../TestRuns/FilterField";
 import {
   TimeFilterContext,
   TimeFilterOptions,
   withinTeamRetentionLimit,
-} from "../TimeFilterContextRoot";
-import { TestContext } from "./TestContextRoot";
+} from "../../TimeFilterContextRoot";
+import { TestContext } from "../TestContextRoot";
 import { TestList } from "./TestList";
 
 const timeFilterLabel: Record<TimeFilterOptions, string> = {
